drone control done! Now need to connect with camera and implement the live feeding on the control screen for Manual Control implement
