+++ app_AS_only/app/src/main/res/layout/activity_control_selection.xml
@@ -4,36 +4,27 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".ControlSelection">
 
     <Button
-        android:id="@+id/log"
+        android:id="@+id/AutomaticControl"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="log"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/fly"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/fly"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="fly"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:text="Automatic Control"
+        app:layout_constraintBottom_toTopOf="@+id/ManualControl"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.74" />
 
     <Button
-        android:id="@+id/help"
+        android:id="@+id/ManualControl"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="help"
+        android:text="Manual Control"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/fly"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
